Should work now

+++ include/freqAnalysis.h
@@ -15,7 +15,8 @@ typedef struct{
 } freqAnaliser;
 
 freqAnaliser initAnaliser(float freq);
-void processSet(freqAnaliser * a, uint16_t resetLenght);
+void processSet(freqAnaliser * a, float resetLenght);
 float getScoreSquare(freqAnaliser * a);
 void resetScore(freqAnaliser * a);
 #endif
+
